apps/data-science/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
distro==1.9.0
gensim==4.3.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
joblib==1.3.2
numpy==1.26.4
openai==1.13.3
pandas==2.2.0
psycopg2==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smart-open==7.0.1
sniffio==1.3.1
SQLAlchemy==2.0.27
threadpoolctl==3.3.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
wrapt==1.16.0
